Mailer-Service/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mailer-service"
version = "0.1.0"
description = "Mailer Microservice with Amazon SES, JWT verification, and PostgreSQL"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.104.1",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "boto3>=1.28.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]
